Remove old S3 testing fixture
