+++ internal/s3storage/s3setup_test.go
@@ -1,114 +1,105 @@
 package s3storage_test
 
 import (
+	"archive/zip"
 	"bytes"
-	"encoding/base64"
-	"os"
+	"encoding/json"
+	"fmt"
+	"net/http/httptest"
+	"strings"
 	"testing"
+	"time"
 
 	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/awserr"
+	"github.com/aws/aws-sdk-go/aws/credentials"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/s3"
 	"github.com/aws/aws-sdk-go/service/s3/s3iface"
-	"github.com/stretchr/testify/assert"
+	"github.com/johannesboyne/gofakes3"
+	"github.com/johannesboyne/gofakes3/backend/s3mem"
+	"github.com/lunarway/release-manager/internal/artifact"
 )
 
-func SkipIfNoAWS(t *testing.T) {
-	if os.Getenv("AWS_CONFIG_FILE") == "" {
-		t.Skip("AWS not configured to run AWS integration tests")
+// setupS3 instantiates a fake S3 backend and seeds a bucket with provided
+// artifacts.
+func setupS3(t *testing.T, bucket string, artifacts ...artifact.Spec) (s3iface.S3API, func()) {
+	t.Helper()
+	backend := s3mem.New()
+	faker := gofakes3.New(backend)
+	ts := httptest.NewServer(faker.Server())
+
+	// configure S3 client
+	s3Config := &aws.Config{
+		Credentials:      credentials.NewStaticCredentials("YOUR-ACCESSKEYID", "YOUR-SECRETACCESSKEY", ""),
+		Endpoint:         aws.String(ts.URL),
+		Region:           aws.String("eu-central-1"),
+		DisableSSL:       aws.Bool(true),
+		S3ForcePathStyle: aws.Bool(true),
 	}
+	newSession, err := session.NewSession(s3Config)
+	if err != nil {
+		t.Fatalf("Failed to instantiate session: %v", err)
+	}
+
+	s3Client := s3.New(newSession)
+
+	// Create a new bucket using the CreateBucket call.
+	_, err = s3Client.CreateBucket(&s3.CreateBucketInput{
+		Bucket: aws.String(bucket),
+	})
+	if err != nil {
+		t.Fatalf("Failed to create test bucket: %v", err)
+	}
+
+	seedArtifacts(t, s3Client, bucket, artifacts)
+
+	return s3Client, ts.Close
 }
 
-func EnsureTestS3Objects(t *testing.T, setup S3BucketSetup) s3iface.S3API {
-	sess, err := session.NewSession(&aws.Config{
-		Region: aws.String("eu-west-1"),
-	})
-	s3client := s3.New(sess)
-	assert.NoError(t, err)
+func seedArtifacts(t *testing.T, s3Client s3iface.S3API, bucket string, artifacts []artifact.Spec) {
+	t.Helper()
+	for _, artifact := range artifacts {
+		artifactZip := zipArtifact(t, artifact)
 
-	_, err = s3client.CreateBucket(&s3.CreateBucketInput{
-		Bucket: aws.String(setup.BucketName),
-	})
-	aerr, isAwsErr := err.(awserr.Error)
-	if isAwsErr && (aerr.Code() == s3.ErrCodeBucketAlreadyOwnedByYou || aerr.Code() == s3.ErrCodeBucketAlreadyExists) {
-		t.Logf("s3 bucket already exists")
-		err = emptyBucket(t, s3client, setup.BucketName)
-		assert.NoError(t, err)
-	} else {
-		assert.NoError(t, err)
-	}
+		_, err := s3Client.PutObject(&s3.PutObjectInput{
+			Body:   strings.NewReader(artifactZip),
+			Bucket: aws.String(bucket),
 
-	t.Log("s3 bucket ensured")
-	for keyName, object := range setup.Objects {
-
-		objectContent, err := base64.StdEncoding.DecodeString(object.Base64Content)
-		assert.NoError(t, err)
-		_, err = s3client.PutObject(&s3.PutObjectInput{
-			Bucket:   aws.String(setup.BucketName),
-			Key:      aws.String(keyName),
-			Body:     bytes.NewReader(objectContent),
-			Metadata: aws.StringMap(object.Metadata),
+			Key: aws.String(fmt.Sprintf("%s/%s", artifact.Service, artifact.ID)),
 		})
-		assert.NoError(t, err)
-	}
-	return s3client
-}
-
-type S3BucketSetup struct {
-	BucketName string
-	Objects    map[string]S3BucketSetupObject
-}
-
-type S3BucketSetupObject struct {
-	Metadata      map[string]string
-	Base64Content string
-}
-
-func emptyBucket(t *testing.T, s3client *s3.S3, bucket string) error {
-	t.Logf("removing objects from S3 bucket: %s", bucket)
-	for {
-		params := &s3.ListObjectsInput{
-			Bucket: aws.String(bucket),
-		}
-
-		objects, err := s3client.ListObjects(params)
 		if err != nil {
-			return err
+			t.Fatalf("Failed to marshal artifact: %v", err)
 		}
-		//Checks if the bucket is already empty
-		if len((*objects).Contents) == 0 {
-			t.Log("Bucket is already empty")
-			return nil
-		}
-		t.Logf("First object in batch %s", *(objects.Contents[0].Key))
+		t.Logf("Uploaded test artifact %s to bucket", artifact.ID)
 
-		//creating an array of pointers of ObjectIdentifier
-		objectsToDelete := make([]*s3.ObjectIdentifier, 0, 1000)
-		for _, object := range (*objects).Contents {
-			obj := s3.ObjectIdentifier{
-				Key: object.Key,
-			}
-			objectsToDelete = append(objectsToDelete, &obj)
-		}
-		//Creating JSON payload for bulk delete
-		deleteArray := s3.Delete{Objects: objectsToDelete}
-		deleteParams := &s3.DeleteObjectsInput{
-			Bucket: aws.String(bucket),
-			Delete: &deleteArray,
-		}
-		//Running the Bulk delete job (limit 1000)
-		_, err = s3client.DeleteObjects(deleteParams)
-		if err != nil {
-			return err
-		}
-		if *(*objects).IsTruncated { //if there are more objects in the bucket, IsTruncated = true
-			params.Marker = (*deleteParams).Delete.Objects[len((*deleteParams).Delete.Objects)-1].Key
-			t.Logf("Requesting next batch %s", *(params.Marker))
-		} else { //if all objects in the bucket have been cleaned up.
-			break
-		}
+		// sleep to ensure we get different modified timestamps in S3
+		time.Sleep(2 * time.Millisecond)
 	}
-	t.Logf("Emptied S3 bucket %s", bucket)
-	return nil
+}
+
+func zipArtifact(t *testing.T, artifact artifact.Spec) string {
+	t.Helper()
+	artifactJSON, err := json.Marshal(artifact)
+	if err != nil {
+		t.Fatalf("Failed to marshal artifact: %v", err)
+	}
+
+	var zipBytes bytes.Buffer
+	zipWriter := zip.NewWriter(&zipBytes)
+	if err != nil {
+		t.Fatalf("Failed to instantiate zip writer: %v", err)
+	}
+	zipFile, err := zipWriter.Create("artifact.json")
+	if err != nil {
+		t.Fatalf("Failed to create zip file: %v", err)
+	}
+	_, err = zipFile.Write(artifactJSON)
+	if err != nil {
+		t.Fatalf("Failed to write artifact to zip file: %v", err)
+	}
+	err = zipWriter.Close()
+	if err != nil {
+		t.Fatalf("Failed to close zip file: %v", err)
+	}
+	return zipBytes.String()
 }
